Fix tests and add version suffix
