+++ .circleci/config.yml
@@ -97,14 +97,14 @@ jobs:
       - image: elasticsearch:5
         environment:
           ES_JAVA_OPTS: "-Xms750m -Xmx750m" # https://github.com/10up/wp-local-docker/issues/6
-      - image: elasticsearch:6
+      - image: elasticsearch:6.8.13
         environment:
+          http.port: 9202-9300
           ES_JAVA_OPTS: "-Xms750m -Xmx750m" # https://github.com/10up/wp-local-docker/issues/6
-      - image: elasticsearch:7
+      - image: elasticsearch:7.10.1
         environment:
-          ES_JAVA_OPTS: "-Xms750m -Xmx750m" # https://github.com/10up/wp-local-docker/issues/6
-      - image: elasticsearch:8
-        environment:
+          http.port: 9203-9300
+          discovery.type: single-node
           ES_JAVA_OPTS: "-Xms750m -Xmx750m" # https://github.com/10up/wp-local-docker/issues/6
       - image: datadog/docker-dd-agent
         environment:
@@ -166,21 +166,18 @@ jobs:
           command: dockerize -wait tcp://localhost:6379 -timeout 1m
 
       - run:
-          name: Wait for ElasticSearch (1)
+          name: Wait for ElasticSearch (2)
           command: dockerize -wait http://localhost:9200 -timeout 1m
 
       - run:
-          name: Wait for ElasticSearch (2)
+          name: Wait for ElasticSearch (5)
           command: dockerize -wait http://localhost:9201 -timeout 1m
       - run:
-          name: Wait for ElasticSearch (3)
+          name: Wait for ElasticSearch (6)
           command: dockerize -wait http://localhost:9202 -timeout 1m
       - run:
-          name: Wait for ElasticSearch (4)
+          name: Wait for ElasticSearch (7)
           command: dockerize -wait http://localhost:9203 -timeout 1m
-      - run:
-          name: Wait for ElasticSearch (5)
-          command: dockerize -wait http://localhost:9204 -timeout 1m
 
       - run:
           name: Wait for Datadog Agent
@@ -201,7 +198,7 @@ jobs:
       - run:
           name: Testing
           command: |
-                INTEGRATION=1 go test -v -race `go list ./contrib/... | grep -v -e grpc.v12 -e google.golang.org/api`
+                INTEGRATION=1 go test -v -race `go list ./contrib/elastic/... | grep -v -e grpc.v12 -e google.golang.org/api`
                 go get google.golang.org/grpc@v1.29.0 # https://github.com/grpc/grpc-go/issues/3726
                 go test -v ./contrib/google.golang.org/api/...
                 go get google.golang.org/grpc@v1.2.0
