+++ contrib/elastic/go-elasticsearch.v5/elastictrace.go
