+++ contrib/elastic/go-elasticsearch.v5/option.go
